added evaluation for the DRUG task
