+++ src/utils.py
@@ -68,132 +68,6 @@ def summarize(report: str, ailment: str) -> str:
     return summary
 
 
-def evaluate(context: Prompt, test_df: pd.DataFrame, ailment: str, machine, agree_fn: Callable) -> float:
-    """
-    Evaluates the model on the given test data.
-
-    Args:
-        context (Prompt): Current context for the model
-        test_df (pd.DataFrame): Test data
-        ailment (str): The ailment to evaluate the model for
-        machine (Agent): The machine agent
-        agree_fn (Callable): Function to check if the explanation agrees with the prediction
-
-    Returns:
-        float: The overall accuracy of the model
-    """
-    print(f"Evaluating for {ailment}...")
-
-    # initialize the counters
-    total = 0
-    correct_preds = 0
-    correct_expls = 0
-    correct = 0
-
-    # prediction prompt
-    pred_prompt = deepcopy(context)
-    pred_prompt.extend([
-        {
-            "role": "system",
-            "content": """You are a helpful radiology expert, with detailed knowledge of Atelectasis, Pneumonia, Pleural Effusion, Cardiomegaly, Pneumothorax.
-            You have to strictly response in, no extra text:
-            *Prediction: Yes/No*
-            """
-        },
-        {
-            "role": "user",
-            "content": f"Given the following chest XRay, you have to predict the presence of {ailment}."
-        }
-    ]
-    )
-
-    expl_prompt = deepcopy(context)
-    expl_prompt.extend([
-        {
-            # TODO: check prompt
-            "role": "system",
-            "content": """You are a helpful radiology expert, with detailed knowledge of Atelectasis, Pneumonia, Pleural Effusion, Cardiomegaly, Pneumothorax.
-            You have to strictly response in, no extra text:
-            *Prediction: Yes*
-            *Explanation: <Your explanation here>*
-            """
-        },
-        {
-            "role": "user",
-            "content": f"Given the following chest XRay, you have to explain the presence of {ailment}."
-        },
-        {
-            "role": "assistant",
-            "content": "*Prediction: Yes*"
-        }
-    ]
-    )
-
-    # iterate over the test data
-    for _, row in test_df.iterrows():
-        total += 1
-        report = row["report"]
-        label = row["label"]
-        img = encode_image(row["img_path"])
-
-        # get prediction
-        img_msg = {
-            "role": "user",
-            "content": [
-                {
-                    "type": "image_url",
-                    "image_url": {
-                        "url": f"data:image/jpeg;base64,{img}"
-                    }
-                }
-            ]
-        }
-        y = "problem"
-        while y == "problem":
-            y, _, _ = machine.ask(None, pred_prompt + [img_msg], is_prompt=True)
-            if y == "problem":
-                print("Problem with prediction, retrying...")
-
-        prediction = y
-
-        # if label != ailment, we don't care about explanation, only prediction
-        if label != ailment:
-            matchOK = "no" in prediction.lower()
-            correct_preds += 1 if matchOK else 0
-            correct_expls += 1 if matchOK else 0
-            correct += 1 if matchOK else 0
-        else:
-            if "no" in prediction.lower():
-                # the ailment IS present, but the model predicted it as absent
-                pass
-            else:
-                # get explanation
-                y = "problem"
-                while y == "problem":
-                    y, explanation, _ = machine.ask(None, expl_prompt + [img_msg], is_prompt=True)
-                    if y == "problem":
-                        print("Problem with explanation, retrying...")
-
-                # check if the prediction is correct
-                matchOK = "yes" in prediction.lower()
-                agreeOK = agree_fn(explanation, report)
-                correct_preds += 1 if matchOK else 0 # this is kept as a check so that other things don't get matched
-                correct_expls += 1 if agreeOK else 0
-                correct += 1 if matchOK and agreeOK else 0
-
-    # print the results
-    print(f"Ailment: {ailment}")
-    print(f"Total: {total}")
-    print(f"Correct Predictions: {correct_preds}")
-    print(f"Correct Explanations: {correct_expls}")
-    print(f"Correct Overall: {correct}")
-    print(f"Prediction Accuracy: {100*correct_preds / total:.2f}")
-    print(f"Explanation Accuracy: {100*correct_expls / total:.2f}")
-    print(f"Overall Accuracy: {100*correct / total:.2f}")
-
-    return correct / total
-
-
 def evaluate_many(context: Prompt, test_data, machine, agree_fn: Callable) -> None:
     """
     Evaluates the model on multiple test data.
